+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - New `watermelon.recommend` command, to recommend a watermelon to everyone on the repo
 - New `watermelon.login` command, to login using GitHub from your keyboard
+- We now track every time the `watermelon.show` command is called 
 
 ## [1.4.5]